break up offline area monopoly
